Fix lint issues
