+++ packages/frontend/containers/expenses/expense-listing.container.tsx
@@ -3,7 +3,7 @@ import { useCallback, useEffect, useState } from 'react';
 import { initialMonthFilterOptions } from '$blocks/monthly-transaction-list/monthly-transaction-list';
 import { useFirstTransaction } from '$hooks/transaction/useFirstTransaction';
 import { useViewTransitionRouter } from '$hooks/useViewTransitionRouter';
-import { ExpenseListing } from '$pages/expenses/expense-listing';
+import { ExpenseListing } from '$pages/expenses/expense.listing';
 import { parseYearMonthFromString } from '$utils/formatDate';
 
 interface ExpenseListingContainerProps {
