+++ packages/frontend/package.json
@@ -5,10 +5,12 @@
   "dependencies": {
     "@local/types": "*",
     "@reduxjs/toolkit": "^1.9.1",
+    "@testing-library/jest-dom": "^6.1.5",
     "chart.js": "^3.7.1",
     "chartjs-plugin-zoom": "^1.2.1",
     "clsx": "^1.2.1",
     "date-fns": "^2.30.0",
+    "http-proxy-middleware": "^2.0.6",
     "next": "^13.3.0",
     "next-pwa": "^5.6.0",
     "react": "^18.0.0",
@@ -65,7 +67,6 @@
     "@silte/rtk-codegen-openapi": "1.0.2",
     "@tailwindcss/forms": "^0.5.0",
     "@tailwindcss/typography": "^0.5.8",
-    "@testing-library/jest-dom": "^6.1.5",
     "@testing-library/react": "^14.1.2",
     "@testing-library/user-event": "^14.5.1",
     "@types/jest": "^27.4.1",
@@ -85,7 +86,6 @@
     "eslint-plugin-jsx-a11y": "^6.5.1",
     "eslint-plugin-prettier": "^4.0.0",
     "eslint-plugin-react": "^7.29.4",
-    "http-proxy-middleware": "^2.0.6",
     "npm-run-all": "^4.1.5",
     "postcss": "^8.4.12",
     "postcss-cli": "^9.1.0",
