+++ packages/frontend/src/pages/incomes/income.listing.tsx
@@ -1,17 +1,14 @@
-import { useExpensesFindAllByUserQuery } from '$api/generated/financerApi';
+import { useIncomesFindAllByUserQuery } from '$api/generated/financerApi';
 import { LatestTransactions } from '$blocks/latest-transactions/latest-transactions';
 import { initialMonthFilterOptions } from '$blocks/monthly-transaction-list/monthly-transaction-list';
 import { Pager } from '$blocks/pager/pager';
 import { monthNames } from '$constants/months';
 import { ButtonInternal } from '$elements/button/button.internal';
-import { Heading } from '$elements/heading/heading';
 import { Icon, IconName } from '$elements/icon/icon';
-import { LinkList } from '$elements/link-list/link-list';
-import { LinkListLink } from '$elements/link-list/link-list.link';
 import { LoaderSuspense } from '$elements/loader/loader-suspense';
 import { UpdatePageInfo } from '$renderers/seo/updatePageInfo';
 
-interface ExpenseListingProps {
+interface IncomeListingProps {
   filterOptions: typeof initialMonthFilterOptions;
   firstAvailableTransaction: Date;
   initialPageToLoad: number;
@@ -19,27 +16,27 @@ interface ExpenseListingProps {
   onMonthOptionChange: (direction: 'next' | 'previous') => void;
 }
 
-export const ExpenseListing = ({
+export const IncomeListing = ({
   filterOptions,
   firstAvailableTransaction,
   initialPageToLoad,
   onPageChange,
   onMonthOptionChange,
-}: ExpenseListingProps): JSX.Element => {
+}: IncomeListingProps): JSX.Element => {
   const pageVisibleYear = filterOptions.year;
   const pageVisibleMonth = monthNames[filterOptions.month - 1];
 
   return (
     <>
       <UpdatePageInfo
-        title="Expenses"
+        title="Incomes"
         backLink="/statistics"
         headerAction={
           <ButtonInternal
-            link="/statistics/expenses/add"
+            link="/statistics/incomes/add"
             className="inline-flex items-center justify-center -mr-3 h-11 w-11"
           >
-            <span className="sr-only">Add expense</span>
+            <span className="sr-only">Add income</span>
             <Icon type={IconName.plus} />
           </ButtonInternal>
         }
@@ -68,7 +65,7 @@ export const ExpenseListing = ({
         <LatestTransactions
           filterOptions={filterOptions}
           className="mt-4"
-          useDataHook={useExpensesFindAllByUserQuery}
+          useDataHook={useIncomesFindAllByUserQuery}
           onPageChange={onPageChange}
           initialPage={initialPageToLoad}
         />
